requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pytest>=8.0.0

# Opcionales: voz (activa con `pip install -r requirements-voice.txt`)
# faster-whisper>=1.0.0
# sounddevice>=0.5.0
# numpy>=1.24.0

# Opcionales: LLMs reales
# anthropic>=0.40.0
# openai>=1.50.0
# langchain-anthropic>=0.3.0
# langchain-openai>=0.2.0
